+++ src/picai_prep/__init__.py
@@ -14,9 +14,9 @@
 
 
 from picai_prep.data_utils import atomic_file_copy, atomic_image_write
-from picai_prep.dcm2dce import Dicom2DCEConverter
-from picai_prep.dcm2mha import Dicom2MHAConverter
-from picai_prep.mha2nnunet import MHA2nnUNetConverter
+from picai_prep.dcm2dce.converter import Dicom2DCEConverter
+from picai_prep.dcm2mha.converter import Dicom2MHAConverter
+from picai_prep.mha2nnunet.converter import MHA2nnUNetConverter
 from picai_prep.nnunet2nndet import nnunet2nndet
 
 print("If you have questions or suggestions, feel free to open an issue " +
