--- src/picai_prep/dcm2mha/__init__.py
@@ -1,2 +1,0 @@
-from case import Dicom2MHACase
-from converter import Dicom2MHAConverter
