--- src/picai_prep/mha2nnunet/__init__.py
@@ -1,2 +1,0 @@
-from case import MHA2nnUNetCase
-from converter import MHA2nnUNetConverter
