+++ src/picai_prep/mha2nnunet/converter.py
@@ -23,6 +23,7 @@ from picai_prep.data_utils import PathLike
 from picai_prep.preprocessing import PreprocessingSettings
 from picai_prep.utilities import mha2nnunet_schema
 from picai_prep.mha2nnunet.case import MHA2nnUNetCase
+from picai_prep.mha2nnunet.settings import MHA2nnUNetSettings
 
 
 class MHA2nnUNetConverter(Converter):
